Expose getTransactions endpoints
