+++ src/apiCalls/transactions/getTransactionsCount.ts
@@ -1,0 +1,4 @@
+import { getTransactionsBuilder } from './helpers';
+
+const isCount = true;
+export const getTransactions = getTransactionsBuilder(isCount);
